brokengit add .! attempted refactor of code, please do not
