+++ kernel/err.c
@@ -1,10 +1,8 @@
-#include "text.h"
-#include "defs.h"
+#include "include/err.c"
+#include "include/text.h"
 
-#ifndef ERR_H
-#define ERR_H
-
-void msg(enum MSG_TYPE type, enum ERRSIG sig, char* supp) {
+void msg(enum MSG_TYPE type, enum ERRSIG sig, char* supp);
+{
   unsigned char msg_style;
   if (type == INFO) {
     msg_style = COLOUR(BLACK, B_CYAN);
@@ -27,5 +25,3 @@ void msg(enum MSG_TYPE type, enum ERRSIG sig, char* supp) {
     free(sigbuf);
   }
 }
-
-#endif
