+++ kernel/kernel.c
@@ -1,16 +1,15 @@
-#include "text.h"
-#include "ata.h"
-#include "config.h"
-#include "defs.h"
-#include "util.h"
-#include "pic.h"
-#include "ih.h"
+#include "include/text.h"
+#include "include/ata.h"
+#include "include/config.h"
+#include "include/util.h"
+#include "include/pic.h"
+#include "include/ih.h"
 #include "kapps/kshell.h"
-#include "fs.h"
-#include "hwinf.h"
-#include "timer.h"
-#include "memring.h"
-#include "err.h"
+#include "include/fs.h"
+#include "include/hwinf.h"
+#include "include/timer.h"
+#include "include/memring.h"
+#include "include/err.h"
 
 struct idt_entry {
   unsigned short offset_low; // low 16 bits of offset
