+++ kernel/memring.c
@@ -1,23 +1,6 @@
-#include "text.h"
-#include "util.h"
-
-#ifndef MEMRING_H
-#define MEMRING_H
-
-#define MEMBLK_SIZE 16
-#define MEMRING_LOC 0x12000
-#define MEM_LOC (MEMRING_LOC * MEMBLK_SIZE)
-#define MEMRING_LEN 1024 // 16k of mem free. don't try and increase this, youll prob wind up overwriting the os itself
-
-#define FREE 0x00
-#define BLK_START 0x01 // 0b01
-#define IN_USE 0x02 // 0b10
-#define BLK_END 0x04 // 0b100
-
-// the memring sits at 0x12_000 whereas the memory it points to sits at 0x120_000 (<< 4)
-// each byte in the memring points to a 16 byte block
-
-unsigned char * memring = (unsigned char *) MEMRING_LOC;
+#include "include/memring.h"
+#include "include/text.h"
+#include "include/util.h"
 
 void mem_init() {
   memzero(MEMRING_LOC, MEMRING_LEN * MEMBLK_SIZE);
@@ -103,5 +86,3 @@ void * realloc(void *ptr, unsigned int len) {
     }
   }
 }
-
-#endif
