+++ kernel/timer.c
@@ -1,16 +1,5 @@
-#include "util.h"
-
-#ifndef TIMER_H
-#define TIMER_H
-
-// 1/3 of the ntsc colour burst frequency, or 1/12 of the 315/22 MHz.
-// This is the frequency that the PIT runs at.
-#define PIT_FREQ_HZ 1193182
-
-#define TIMER_DPORT_C0 0x40
-#define TIMER_DPORT_C1 0x41
-#define TIMER_DPORT_C2 0x42
-#define TIMER_CPORT 0x43
+#include "include/timer.h"
+#include "include/util.h"
 
 // get the divisor - the amount the time needs to count down from - from a number in hz
 static inline unsigned short get_divisor(unsigned int freq) {
@@ -32,5 +21,3 @@ void init_timer(unsigned short divisor) {
   idle();
   pbyte_out(0x40, (divisor >> 8) & 0xff);
 }
-
-#endif
