+++ kernel/util.c
@@ -1,32 +1,8 @@
-#include "text.h"
-#include "defs.h"
-// #include "memring.h"
-// cyclic, declare what we want
-
-void *malloc(unsigned int);
-void free(void *);
-
-enum MSG_TYPE;
-void msg(enum MSG_TYPE, enum ERRSIG, char *);
+#include "include/text.h"
+#include "include/memring.h"
 
 // not really using standard library - stdarg just provides platform-dependent defines 
 #include <stdarg.h>
-
-#ifndef UTIL_H
-#define UTIL_H
-
-/*void cp437() {
-  // prints cp437
-
-  write_str_at("  Codepage 437  ", POS((VGA_WIDTH - 16), (VGA_HEIGHT - 16 - 2)), COLOUR(MAGENTA, B_GREEN));
-  for (int cph = 0; cph < 16; ++cph) {
-    write_cell(nybble_to_hex(cph), POS((VGA_WIDTH - 16 - 1), (VGA_HEIGHT - (16 - cph))), COLOUR(RED, WHITE));
-    write_cell(nybble_to_hex(cph), POS((VGA_WIDTH - (16 - cph)), (VGA_HEIGHT - 16 - 1)), COLOUR(RED, WHITE));
-    for (int cpw = 0; cpw < 16; ++cpw) {
-      write_cell((cph * 16) + cpw, POS((VGA_WIDTH - (16 - cpw)), (VGA_HEIGHT - (16 - cph))), COLOUR(YELLOW, B_CYAN));
-    }
-  }
-}*/
 
 int strlen(char *str) {
   int i = -1;
@@ -195,9 +171,6 @@ static inline char *strcat(char *out, char *in) {
   return out;
 }
 
-#define NOT_WAITING 0x00
-#define HAS_INTEGER 0x10
-#define AFTER_PERCENT -1
 void sprintf(char *dest, const char *fmt, ...) {
   va_list args;
   va_start(args, fmt); // second parameter is the last arg before variadic
@@ -290,7 +263,6 @@ void to_ver_string(struct far_ver * ver, char * vbuf) {
   sprintf(vbuf, "FarOS v%d.%d.%d", ver -> major, ver -> minor, ver -> patch);
 }
 
-#define trace_ch_until(a,b) trace_ch_until_with(a,b,0)
 unsigned short trace_ch_until_with(char *str, int until, int start) {
   unsigned short int runx = start;
   for (int ii = 0; ii < until; ii++) {
@@ -314,5 +286,3 @@ unsigned short trace_ch_until_with(char *str, int until, int start) {
   }
   return runx;
 }
-
-#endif
